🔄 [MOD] (ilorm) Ilorm work now with an internal class instantiate as singleton
