+++ packages/ilorm-connector-mongo/test/readData.test.js
@@ -8,7 +8,7 @@ const ilormMongo = require('../index');
 
 const DB_URL = 'mongodb://localhost:27017/ilorm';
 
-const { Schema, declareModel, newModel, } = ilorm;
+const { Schema, } = ilorm;
 
 describe('ilorm-connector-mongodb', () => {
   describe('test/readData', () => {
@@ -66,9 +66,9 @@ describe('ilorm-connector-mongodb', () => {
     });
 
     it('Could query Guillaume', async() => {
-      class User extends newModel(modelFactoryParams) {}
+      class User extends ilorm.newModel(modelFactoryParams) {}
 
-      declareModel(User);
+      ilorm.declareModel(User);
 
       const user = await User.query()
         .firstName.is('Guillaume')
@@ -80,9 +80,9 @@ describe('ilorm-connector-mongodb', () => {
     });
 
     it('Could query two users', async() => {
-      class User extends newModel(modelFactoryParams) {}
+      class User extends ilorm.newModel(modelFactoryParams) {}
 
-      declareModel(User);
+      ilorm.declareModel(User);
 
       const users = await User.query()
         .gender.is('M')
