--- packages/ilorm/lib/model/models.map.js
@@ -1,6 +1,0 @@
-'use strict';
-
-// Map used to store every declared model to reference it, in another schema:
-const classIndex = new Map();
-
-module.exports = classIndex;
