+++ packages/ilorm/lib/query/baseQuery.classFactory/stream.method.js
