+++ packages/ilorm/lib/query/index.js
@@ -1,3 +1,9 @@
 'use strict';
 
-module.exports = require('./query.class');
+const baseQueryClassFactory = require('./baseQuery.classFactory');
+const queryFactory = require('./query.factory');
+
+module.exports = {
+  baseQueryClassFactory,
+  queryFactory,
+};
