+++ packages/ilorm/lib/query/query.factory.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const { getQuery, } = require('./query.class');
 const { FIELDS, SELECT_BEHAVIOR, } = require('ilorm-constants').QUERY;
 const { SCHEMA, MODEL, CONNECTOR, QUERY, UPDATE, SORT, SELECT, } = FIELDS;
 
@@ -27,17 +26,19 @@ const selectBaseFactory = () => ({
 
 /**
  * Create a new Query class from the given context
+ * @param {Ilorm} ilorm The ilorm context
  * @param {Model} model The model associated with the Query
  * @returns {Query} A new Query Class
  */
-const queryFactory = ({ model, }) => {
+const queryFactory = ({ ilorm, model, }) => {
+  const { BaseQuery, } = ilorm;
   const connector = model.getConnector();
   const schema = model.getSchema();
 
   /**
    * The InternalQuery is a class created dynamically in function of the model.
    */
-  class InternalQuery extends getQuery() {}
+  class InternalQuery extends BaseQuery {}
 
   Object.defineProperties(InternalQuery.prototype, {
     [SCHEMA]: defineProperty(schema),
